workers/asr-worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
whisperx==3.1.1
torch==2.1.1
torchaudio==2.1.1
transformers==4.36.2
datasets==2.16.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
opencv-python==4.8.1.78
Pillow==10.1.0
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
